Make preliminary dashboard show
