--- WalletWasabi.Fluent/FakeTile.cs
@@ -1,8 +1,0 @@
-namespace WalletWasabi.Fluent;
-
-public class FakeTile
-{
-	public string Value { get; set; }
-	public string Title { get; set; }
-	public string Icon { get; set; }
-}
