+++ WalletWasabi.Fluent/ViewModels/Wallets/Home/Tiles/Design/FakeTileViewModel.cs
@@ -1,0 +1,8 @@
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles.Design;
+
+public class FakeTileViewModel : ViewModelBase
+{
+	public string Value { get; set; }
+	public string Title { get; set; }
+	public string Icon { get; set; }
+}
