+++ WalletWasabi.Fluent/ViewModels/Wallets/Home/Tiles/Design/WalletDashboardViewModelDesign.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles.Design;
+
+internal class WalletDashboardViewModelDesign : IWalletDashboardViewModel
+{
+	public ICollection<ViewModelBase> Children { get; set; } = new List<ViewModelBase>();
+}
