+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/Design/FakeTileView.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Home.Tiles.Design;
+public partial class FakeTileView : UserControl
+{
+	public FakeTileView()
+	{
+		InitializeComponent();
+	}
+
+	private void InitializeComponent()
+	{
+		AvaloniaXamlLoader.Load(this);
+	}
+}
