+++ WalletWasabi.Fluent/Views/Wallets/WalletView.axaml
@@ -2,8 +2,6 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
-             xmlns:behaviors="using:WalletWasabi.Fluent.Behaviors"
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
              xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets"
              xmlns:home="using:WalletWasabi.Fluent.Views.Wallets.Home"
@@ -71,8 +69,9 @@
           </StackPanel>
           <TextBlock TextWrapping="NoWrap" TextTrimming="CharacterEllipsis" Text="{Binding WalletName, FallbackValue=My Wallet with a very long name}" />
         </DockPanel>
-        <Separator DockPanel.Dock="Bottom" Margin="-200 13 -200 0" HorizontalAlignment="Stretch" />
-        <home:WalletDashboard />
+        <Separator DockPanel.Dock="Bottom" Margin="-200 13 -200 10" HorizontalAlignment="Stretch" />
+        <home:WalletDashboard DataContext="{Binding WalletDashboard}" />
+        <Separator DockPanel.Dock="Bottom" Margin="-200 13 -200 10" HorizontalAlignment="Stretch" />
       </StackPanel>
     </c:ContentArea.Title>
     <DockPanel Name="Panel" VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
